Move TOCContext to own file
