+++ packages/core/src/components/context/TOCContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react'
+
+export const TOCContext = createContext({
+  visibleTOCElements: [] as string[],
+  isInSidebar: false
+})
