+++ packages/core/src/components/shared/TableOfContents.tsx
@@ -12,17 +12,17 @@ import { withSectionContext } from '../context/SectionContext'
 import { TOCQuestion } from './TOCQuestion'
 import { TOCSection } from './TOCSection'
 
-export const renderChildNodes = createRenderChildNodes({
-  section: withSectionContext(TOCSection),
-  question: withQuestionContext(TOCQuestion)
-})
-
 export const mkTableOfContents = (options: {
   showAttachmentLinks: boolean
   showAnsweringInstructions: boolean
   isInSidebar?: boolean
 }) => {
   const { showAnsweringInstructions, isInSidebar } = options
+
+  const renderChildNodes = createRenderChildNodes({
+    section: withSectionContext(TOCSection),
+    question: withQuestionContext(TOCQuestion)
+  })
 
   const TableOfContents: React.FunctionComponent<ExamComponentProps> = () => {
     const { root, maxScore } = useContext(CommonExamContext)
